Only worry about Genesis City
