+++ Explorer/Assets/DCL/Landscape/TerrainGenerator.cs
@@ -1,5 +1,6 @@
 using Cysharp.Threading.Tasks;
 using DCL.Diagnostics;
+using DCL.Landscape.Config;
 using DCL.Landscape.Jobs;
 using DCL.Landscape.NoiseGeneration;
 using DCL.Landscape.Settings;
@@ -11,6 +12,10 @@ using System.Runtime.InteropServices;
 using System.Threading;
 using DCL.Profiling;
 using DCL.Utilities;
+using GPUInstancerPro;
+using System.IO;
+using System.Linq;
+using System.Text;
 using Unity.Collections;
 using Unity.Collections.LowLevel.Unsafe;
 using Unity.Jobs;
@@ -221,6 +226,9 @@ namespace DCL.Landscape
                     {
                         rootGo = factory.InstantiateSingletonTerrainRoot(TERRAIN_OBJECT_NAME);
                         rootGo.position = new Vector3(0, ROOT_VERTICAL_SHIFT, 0);
+
+                        if (LandscapeData.LOAD_TREES_FROM_STREAMINGASSETS)
+                            await LoadTreesAsync(terrainGenData.treeAssets, rootGo);
 
                         Ocean = factory.CreateOcean(rootGo);
                         Wind = factory.CreateWind();
@@ -564,5 +572,99 @@ namespace DCL.Landscape
 
             noiseGenCache.Dispose();
         }
+
+        private static string TreeFilePath => $"{Application.streamingAssetsPath}/Trees.bin";
+
+        private static void SaveTrees(ChunkModel[] chunks)
+        {
+            TreePrototype[] treePrototypes = chunks[0].TerrainData.treePrototypes;
+            var treeTransforms = new List<Matrix4x4>[treePrototypes.Length];
+
+            for (int i = 0; i < treeTransforms.Length; i++)
+                treeTransforms[i] = new List<Matrix4x4>();
+
+            foreach (ChunkModel chunk in chunks)
+            {
+                Vector3 terrainPosition = chunk.terrain.GetPosition();
+                Vector3 terrainSize = chunk.TerrainData.size;
+
+                foreach (TreeInstance tree in chunk.TerrainData.treeInstances)
+                {
+                    Vector3 position = Vector3.Scale(tree.position, terrainSize) + terrainPosition;
+                    Quaternion rotation = Quaternion.Euler(0f, tree.rotation * Mathf.Rad2Deg, 0f);
+                    Vector3 scale = new Vector3(tree.widthScale, tree.heightScale, tree.widthScale);
+                    treeTransforms[tree.prototypeIndex].Add(Matrix4x4.TRS(position, rotation, scale));
+                }
+            }
+
+            using var stream = new FileStream(TreeFilePath, FileMode.Create, FileAccess.Write);
+            using var writer = new BinaryWriter(stream, new UTF8Encoding(false));
+
+            // To help us allocate the correct size buffer when loading.
+            writer.Write(treeTransforms.Max(i => i.Count));
+
+            foreach (List<Matrix4x4> transforms in treeTransforms)
+            {
+                writer.Write(transforms.Count);
+
+                foreach (Matrix4x4 transform in transforms)
+                    for (int i = 0; i < 16; i++)
+                        writer.Write(transform[i]);
+            }
+        }
+
+#if GPUI_PRO_PRESENT
+        private static async UniTask LoadTreesAsync(LandscapeAsset[] treePrototypes, Transform terrainRoot)
+        {
+            const int SIZE_OF_MATRIX = 64;
+
+            unsafe
+            {
+                if (SIZE_OF_MATRIX != sizeof(Matrix4x4))
+                    throw new Exception("The size of the Matrix4x4 struct is wrong");
+            }
+
+            var rendererKeys = new int[treePrototypes.Length];
+
+            for (int prototypeIndex = 0; prototypeIndex < treePrototypes.Length; prototypeIndex++)
+                GPUICoreAPI.RegisterRenderer(terrainRoot, treePrototypes[prototypeIndex].asset,
+                    out rendererKeys[prototypeIndex]);
+
+            await using var stream = new FileStream(TreeFilePath, FileMode.Open, FileAccess.Read,
+                FileShare.Read);
+
+            using var reader = new BinaryReader(stream, new UTF8Encoding(false));
+
+            int maxInstanceCount = reader.ReadInt32();
+            var buffer = new Matrix4x4[maxInstanceCount];
+
+            for (int prototypeIndex = 0; prototypeIndex < treePrototypes.Length; prototypeIndex++)
+            {
+                int instanceCount = reader.ReadInt32();
+
+                unsafe
+                {
+                    fixed (Matrix4x4* bufferPtr = buffer)
+                        ReadReliably(reader, new Span<byte>(bufferPtr, instanceCount * SIZE_OF_MATRIX));
+                }
+
+                GPUICoreAPI.SetTransformBufferData(rendererKeys[prototypeIndex], buffer,
+                    count: instanceCount);
+            }
+        }
+
+        private static void ReadReliably(BinaryReader reader, Span<byte> buffer)
+        {
+            while (buffer.Length > 0)
+            {
+                int read = reader.Read(buffer);
+
+                if (read <= 0)
+                    throw new EndOfStreamException("Read zero bytes");
+
+                buffer = buffer.Slice(read);
+            }
+        }
+#endif
     }
 }
